Move nested types out of Channel
