+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/ChannelExtensions.cs
@@ -26,7 +26,7 @@ namespace System.Threading.Tasks.Channels
 
             return (IObservable<T>)s_channelToObservable.GetValue(
                 source, 
-                s => new Channel.ChannelObservable<T>((IReadableChannel<T>)s));
+                s => new ChannelObservable<T>((IReadableChannel<T>)s));
         }
 
         /// <summary>Creates an observer for a writeable channel.</summary>
@@ -38,7 +38,7 @@ namespace System.Threading.Tasks.Channels
             if (target == null)
                 throw new ArgumentNullException(nameof(target));
 
-            return new Channel.ChannelObserver<T>(target);
+            return new ChannelObserver<T>(target);
         }
 
         /// <summary>Gets an awaiter that enables directly awaiting a channel to read data from it.</summary>
@@ -65,7 +65,7 @@ namespace System.Threading.Tasks.Channels
             if (channel == null)
                 throw new ArgumentNullException(nameof(channel));
 
-            return new Channel.AsyncEnumerator<T>(channel, cancellationToken);
+            return new AsyncEnumerator<T>(channel, cancellationToken);
         }
 
         /// <summary>Creates a case-select builder and adds a case for channel reading.</summary>
@@ -73,14 +73,14 @@ namespace System.Threading.Tasks.Channels
         /// <param name="channel">The channel from which to read.</param>
         /// <param name="action">The action to invoke with data read from the channel.</param>
         /// <returns>This builder.</returns>
-        public static Channel.CaseBuilder CaseRead<T>(IReadableChannel<T> channel, Action<T> action) => new Channel.CaseBuilder().CaseRead(channel, action);
+        public static CaseBuilder CaseRead<T>(IReadableChannel<T> channel, Action<T> action) => new CaseBuilder().CaseRead(channel, action);
 
         /// <summary>Creates a case-select builder and adds a case for channel reading.</summary>
         /// <typeparam name="T">Specifies the type of data in the channel.</typeparam>
         /// <param name="channel">The channel from which to read.</param>
         /// <param name="func">The asynchronous function to invoke with data read from the channel.</param>
         /// <returns>This builder.</returns>
-        public static Channel.CaseBuilder CaseRead<T>(IReadableChannel<T> channel, Func<T, Task> func) => new Channel.CaseBuilder().CaseRead(channel, func);
+        public static CaseBuilder CaseRead<T>(IReadableChannel<T> channel, Func<T, Task> func) => new CaseBuilder().CaseRead(channel, func);
 
         /// <summary>Creates a case-select builder and adds a case for channel writing.</summary>
         /// <typeparam name="T">Specifies the type of data in the channel.</typeparam>
@@ -88,7 +88,7 @@ namespace System.Threading.Tasks.Channels
         /// <param name="item">The data to write to the channel</param>
         /// <param name="action">The action to invoke after the data has been written.</param>
         /// <returns>This builder.</returns>
-        public static Channel.CaseBuilder CaseWrite<T>(IWritableChannel<T> channel, T item, Action action) => new Channel.CaseBuilder().CaseWrite(channel, item, action);
+        public static CaseBuilder CaseWrite<T>(IWritableChannel<T> channel, T item, Action action) => new CaseBuilder().CaseWrite(channel, item, action);
 
         /// <summary>Creates a case-select builder and adds a case for channel writing.</summary>
         /// <typeparam name="T">Specifies the type of data in the channel.</typeparam>
@@ -96,7 +96,7 @@ namespace System.Threading.Tasks.Channels
         /// <param name="item">The data to write to the channel</param>
         /// <param name="func">The asynchronous function to invoke after the data has been written.</param>
         /// <returns>This builder.</returns>
-        public static Channel.CaseBuilder CaseWrite<T>(IWritableChannel<T> channel, T item, Func<Task> func) => new Channel.CaseBuilder().CaseWrite(channel, item, func);
+        public static CaseBuilder CaseWrite<T>(IWritableChannel<T> channel, T item, Func<Task> func) => new CaseBuilder().CaseWrite(channel, item, func);
 
         /// <summary>Mark the channel as being complete, meaning no more items will be written to it.</summary>
         /// <param name="channel">The channel to mark as complete.</param>
@@ -108,6 +108,171 @@ namespace System.Threading.Tasks.Channels
                 throw new ArgumentNullException(nameof(channel));
             if (!channel.TryComplete(error))
                 throw ChannelUtilities.CreateInvalidCompletionException();
+        }
+
+        /// <summary>Provides an observer for a writeable channel.</summary>
+        internal sealed class ChannelObserver<T> : IObserver<T>
+        {
+            private readonly IWritableChannel<T> _channel;
+
+            internal ChannelObserver(IWritableChannel<T> channel)
+            {
+                Debug.Assert(channel != null);
+                _channel = channel;
+            }
+
+            public void OnCompleted() => _channel.Complete();
+
+            public void OnError(Exception error) => _channel.Complete(error);
+
+            public void OnNext(T value) => _channel.WriteAsync(value).GetAwaiter().GetResult();
+        }
+
+        /// <summary>Provides an observable for a readable channel.</summary>
+        internal sealed class ChannelObservable<T> : IObservable<T>
+        {
+            private readonly List<IObserver<T>> _observers = new List<IObserver<T>>();
+            private readonly IReadableChannel<T> _channel;
+            private CancellationTokenSource _cancelCurrentOperation;
+
+            internal ChannelObservable(IReadableChannel<T> channel)
+            {
+                Debug.Assert(channel != null);
+                _channel = channel;
+            }
+
+            public IDisposable Subscribe(IObserver<T> observer)
+            {
+                if (observer == null)
+                    throw new ArgumentNullException(nameof(observer));
+
+                lock (_observers)
+                {
+                    _observers.Add(observer);
+                    if (_cancelCurrentOperation == null)
+                    {
+                        _cancelCurrentOperation = new CancellationTokenSource();
+                        CancellationToken token = _cancelCurrentOperation.Token;
+                        Task.Run(() => ForwardAsync(token));
+                    }
+                    return new Unsubscribe(this, observer);
+                }
+            }
+
+            private async void ForwardAsync(CancellationToken cancellationToken)
+            {
+                Exception error = null;
+                try
+                {
+                    IAsyncEnumerator<T> e = _channel.GetAsyncEnumerator(cancellationToken);
+                    while (await e.MoveNextAsync().ConfigureAwait(false))
+                    {
+                        lock (_observers)
+                        {
+                            for (int i = 0; i < _observers.Count; i++)
+                            {
+                                _observers[i].OnNext(e.Current);
+                            }
+                        }
+                    }
+                    await _channel.Completion.ConfigureAwait(false);
+                }
+                catch (Exception exc)
+                {
+                    error = exc;
+                }
+                finally
+                {
+                    lock (_observers)
+                    {
+                        for (int i = 0; i < _observers.Count; i++)
+                        {
+                            if (error != null)
+                                _observers[i].OnError(error);
+                            else
+                                _observers[i].OnCompleted();
+                        }
+                    }
+                }
+            }
+
+            private sealed class Unsubscribe : IDisposable
+            {
+                private readonly ChannelObservable<T> _observable;
+                private IObserver<T> _observer;
+
+                internal Unsubscribe(ChannelObservable<T> observable, IObserver<T> observer)
+                {
+                    Debug.Assert(observable != null);
+                    Debug.Assert(observer != null);
+
+                    _observable = observable;
+                    _observer = observer;
+                }
+
+                public void Dispose()
+                {
+                    lock (_observable._observers)
+                    {
+                        if (_observer != null)
+                        {
+                            bool removed = _observable._observers.Remove(_observer);
+                            Debug.Assert(removed);
+                            _observer = null;
+
+                            if (_observable._observers.Count == 0)
+                            {
+                                _observable._cancelCurrentOperation.Cancel();
+                                _observable._cancelCurrentOperation = null;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        /// <summary>Provides an async enumerator for the data in a channel.</summary>
+        internal sealed class AsyncEnumerator<T> : IAsyncEnumerator<T>
+        {
+            /// <summary>The channel being enumerated.</summary>
+            private readonly IReadableChannel<T> _channel;
+            /// <summary>Cancellation token used to cancel the enumeration.</summary>
+            private readonly CancellationToken _cancellationToken;
+            /// <summary>The current element of the enumeration.</summary>
+            private T _current;
+
+            internal AsyncEnumerator(IReadableChannel<T> channel, CancellationToken cancellationToken)
+            {
+                _channel = channel;
+                _cancellationToken = cancellationToken;
+            }
+
+            public T Current => _current;
+
+            public Task<bool> MoveNextAsync()
+            {
+                ValueTask<T> result = _channel.ReadAsync(_cancellationToken);
+
+                if (result.IsCompletedSuccessfully)
+                {
+                    _current = result.Result;
+                    return ChannelUtilities.TrueTask;
+                }
+
+                return result.AsTask().ContinueWith((t, s) =>
+                {
+                    AsyncEnumerator<T> thisRef = (AsyncEnumerator<T>)s;
+                    try
+                    {
+                        thisRef._current = t.GetAwaiter().GetResult();
+                        return true;
+                    }
+                    catch (ClosedChannelException)
+                    {
+                        return false;
+                    }
+                }, this, CancellationToken.None, TaskContinuationOptions.ExecuteSynchronously, TaskScheduler.Default);
+            }
         }
     }
 }
